Fix insertions bug and remove GpC
Insertions are now removed when Record struct is created, as they
are not needed later on. This saves on computation later and a
little bit of memory.

GpCs are also not used, even if the data is NOMe-seq, so we can
just skip them and never process them to save on some more memory.